feat: 動画登録ページのUIを改修
Fixes #460
