+++ src/app/editor/nicovideo/OriginalSource.tsx
@@ -2,18 +2,18 @@
 import "client-only";
 
 import clsx from "clsx";
-import Image from "next/image";
 import React, { useEffect } from "react";
 
+import { CoolImage } from "~/components/common/CoolImage";
 import { FragmentType, graphql, useFragment } from "~/gql";
 
-import { TagButton } from "../TagButton";
 import {
   useSetSourceId,
   useSetThumbnailUrl,
   useSetTitle,
   useToggleSemitag,
-} from "./Context";
+} from "./RegisterContext";
+import { TagButton } from "./TagButton";
 
 export const Fragment = graphql(`
   fragment RegisterNicovideoPage_OriginalSource on NicovideoOriginalSource {
@@ -25,6 +25,7 @@ export const Fragment = graphql(`
       searchTags(input: { limit: 3 }) {
         items {
           tag {
+            ...RegisterNicovideoPage_TagButton
             id
           }
         }
@@ -53,47 +54,31 @@ export const Original: React.FC<{
   );
 
   return (
-    <div className={clsx(className, ["flex", "gap-x-4"])}>
-      <div
-        className={clsx(
-          ["flex-shrink-0"],
-          ["w-72"],
-          ["flex", "flex-col", "gap-y-4"]
-        )}
-      >
-        <div className={clsx(["flex", "flex-col", "gap-y-1"])}>
-          <div className={clsx(["text-xs"])}>タイトル</div>
-          <div className={clsx(["text-sm", "font-bold"])}>{fragment.title}</div>
-        </div>
-        <div className={clsx(["flex", "flex-col", "gap-y-1"])}>
-          <div className={clsx(["text-xs"])}>サムネイル</div>
-          <div className={clsx()}>
-            <Image
-              className={clsx(["object-scale-down"], ["w-48"])}
-              src={fragment.thumbnailUrl}
-              width={260}
-              height={200}
-              alt={`${fragment.sourceId}のサムネイル`}
-            />
-          </div>
-        </div>
+    <div className={clsx(className, ["flex", "gap-x-8"])}>
+      <div className={clsx(["flex-shrink-0"], ["flex", "flex-col", "gap-y-4"])}>
+        <CoolImage
+          className={clsx(["w-56"], ["h-40"])}
+          src={fragment.thumbnailUrl}
+          width={208}
+          height={160}
+          alt={`${fragment.sourceId}のサムネイル`}
+        />
       </div>
-      <div className={clsx(["flex-grow"], ["flex", "flex-col"])}>
-        <div className={clsx(["text-xs"])}>タグ</div>
+      <div className={clsx(["flex-grow"], ["flex", "flex-col"], ["gap-y-4"])}>
+        <div className={clsx(["text-sm", "font-bold"])}>{fragment.title}</div>
         <div
-          className={clsx(
-            ["mt-1"],
-            ["grid", ["grid-cols-3"], ["gap-x-2"], ["gap-y-2"]]
-          )}
+          className={clsx(["grid", "grid-cols-1", ["gap-x-1"], ["gap-y-1"]])}
         >
           {fragment.tags.map((tag) => (
-            <div
-              key={tag.name}
-              className={clsx(["flex", "flex-col", "gap-y-1"])}
-            >
+            <div key={tag.name} className={clsx(["flex"])}>
               <button
                 type="button"
-                className={clsx(["text-sm", "text-left"])}
+                className={clsx([
+                  "text-xs",
+                  "text-left",
+                  "text-slate-700",
+                  "font-bold",
+                ])}
                 onClick={(e) => {
                   e.preventDefault();
                   toggleSemitag(tag.name);
@@ -102,15 +87,20 @@ export const Original: React.FC<{
                 {tag.name}
               </button>
               {tag.searchTags.items.length === 0 && (
-                <div className={clsx(["text-xs", "select-none"])}>候補なし</div>
+                <div
+                  className={clsx(
+                    ["ml-2"],
+                    ["text-xs", "select-none", "text-slate-500"]
+                  )}
+                >
+                  候補なし
+                </div>
               )}
               <div
-                className={clsx([
-                  "flex",
-                  "flex-col",
-                  "items-start",
-                  "gap-y-0.5",
-                ])}
+                className={clsx(
+                  ["ml-2"],
+                  ["flex", "flex-wrap", "items-start", "gap-x-1", "gap-y-0.5"]
+                )}
               >
                 {tag.searchTags.items.map((item) => (
                   <TagButton key={item.tag.id} tagId={item.tag.id} />
