+++ src/app/editor/nicovideo/RequestNotExists.tsx
@@ -1,7 +1,7 @@
 import clsx from "clsx";
 import { useEffect } from "react";
 
-import { useSetRequestId } from "./Context";
+import { useSetRequestId } from "./RequestContext";
 
 export const NotExists: React.FC = () => {
   const setRequestId = useSetRequestId();
