+++ src/app/editor/nicovideo/TagButton.tsx
@@ -4,14 +4,20 @@ import React from "react";
 import { useQuery } from "urql";
 
 import { CommonTag } from "~/components/CommonTag";
-import { graphql } from "~/gql";
+import { graphql, useFragment } from "~/gql";
 
-import { useToggleTag } from "./Original/Context";
+import { useToggleTag } from "./RegisterContext";
 
+export const Fragment = graphql(`
+  fragment RegisterNicovideoPage_TagButton on Tag {
+    ...CommonTag
+    id
+  }
+`);
 export const Query = graphql(`
   query RegisterNicovideoPage_RegisterForm_Confirm_Tag($id: ID!) {
     getTag(id: $id) {
-      ...CommonTag
+      ...RegisterNicovideoPage_TagButton
       id
     }
   }
@@ -27,19 +33,21 @@ export const TagButton: React.FC<{
   });
   const toggleTag = useToggleTag();
 
-  if (!data) return null;
+  const fragment = useFragment(Fragment, data?.getTag);
+
+  if (!fragment) return null;
 
   return (
     <button
       className={clsx(className, ["flex"], ["text-left"])}
       type="button"
       onClick={() => {
-        toggleTag(data.getTag.id);
+        toggleTag(fragment.id);
       }}
     >
       <CommonTag
         className={clsx(["text-xs"], ["px-1"], ["py-0.5"])}
-        fragment={data.getTag}
+        fragment={fragment}
       />
     </button>
   );
