+++ src/app/request/nicovideo/RequestForm.tsx
@@ -6,8 +6,8 @@ import React, { CSSProperties } from "react";
 import { useFieldArray, useForm } from "react-hook-form";
 import * as z from "zod";
 
-import { RegisterContext } from "~/app/editor/nicovideo/_components/Original/Context";
-import { TagButton } from "~/app/editor/nicovideo/_components/TagButton";
+import { RegisterContext } from "~/app/editor/nicovideo/RegisterContext";
+import { TagButton } from "~/app/editor/nicovideo/TagButton";
 import { LinkVideo } from "~/app/videos/[serial]/Link";
 import { BlueButton } from "~/components/common/Button";
 import OptionalSemitagTagSearcher from "~/components/OptionalSemitagTagSearcher";
